+++ src/main/java/zenpad/misc/manager/NoteSyncManager.java
@@ -1,4 +1,4 @@
-package zenpad.misc;
+package zenpad.misc.manager;
 
 import javax.swing.JTabbedPane;
